Got the OO pygame player working in a basic form.
It sets the cursor to look the same as the one I see on
asciimator.net, so that's neat.  Generally, the code seems tidier
than the non-OO code, so that's neat too.  I still need to
implement the cool playback features of the old player, and I'd
like to get a less stupid way of setting the font size when the
window is resized.