Implemented file opening while running.
If no filename is given on the command line, a default single frame
ASCIImation is shown saying that no file is loaded.  If a filename is
given on the command line, is it played as usual.

To choose a file while the program is running, simply click the open
button.  A file chooser dialog appears, with a default filter for
JSON files.  When a file is chosen, the ASCIImation begins playing
immediately.  No error checking is done, so opening a bad file would
probably raise a neat exception.

FIXED: zero division when playing a single frame ASCIImation.