ASCIIPlaybacks are constructed from ASCIImations
There's no way to just pass a filename to an ASCIIPlayback anymore.
Instead, pass it an ASCIImation with the filename.  More typing, yes,
but it simplifies the ASCIIPlayback constructor so I like it.

Unrelated change: adding a StackSwitcher to the HeaderBar in ASCIImator
Desktop.  It's not as cluttered as I thought.