Stop using Pango markup for text attributes
Not only is it unnecessary when the entire animation label has a single
style in each frame, but the old code broke when characters that looked
like markup were seen by the parser.  Rather than mess around with
sanitizing ASCIImation frames for inclusion in markup, now they're just
set as text, and the style is set separately.